chore: layout composition improvements
