+++ apps/regen-link/app/[locale]/(main)/layout.tsx
@@ -2,8 +2,8 @@ import { type PropsWithChildren } from 'react';
 
 export default function MainLayout({ children }: PropsWithChildren) {
   return (
-    <div className="min-h-screen bg-background">
-      <main className="container mx-auto px-4 py-8">{children}</main>
-    </div>
+    <main className="container flex-grow flex flex-col h-screen mx-auto px-4 py-8 text-center items-center justify-center">
+      {children}
+    </main>
   );
 }
